ボタン設定画面を別ページに変更
