+++ src/components/MainPage.vue
@@ -2,11 +2,13 @@ nc
 <script lang="ts">
 import SiteSettings from "./SiteSettings.vue";
 import KeyInputHistory from "./KeyInputHistory.vue";
+import ButtonSettings from "./Settings/ButtonSettings.vue";
 
 export default {
   components: {
     SiteSettings,
     KeyInputHistory,
+    ButtonSettings,
   },
   data() {
     return {
@@ -43,6 +45,13 @@ export default {
                 :class="{ contentActive: currentTab === 'KeyInputHistory' }"
                 >キー入力履歴表示</span
               ></a
+            >
+          </li>
+          <li class="nav-item">
+            <a class="nav-link" @click="currentTab = 'ButtonSettings'">
+              <span :class="{ contentActive: currentTab === 'ButtonSettings' }"
+                >ボタン設定
+              </span></a
             >
           </li>
           <li class="nav-item">
